fix(kilo-pass): break circular dependency by moving findActiveKiloPassByCardFingerprint to card-fingerprint-gate
