+++ apps/web/src/lib/stripe/index.ts
@@ -7,7 +7,6 @@ import { db, auto_deleted_at } from '@/lib/drizzle';
 import type { User, PaymentMethod, Organization } from '@kilocode/db/schema';
 import {
   kilo_pass_scheduled_changes,
-  kilo_pass_subscriptions,
   payment_methods,
   kilocode_users,
   auto_top_up_configs,
@@ -217,59 +216,6 @@ export async function isCardFingerprintEligibleForFreeCredits(
       ),
     })) === undefined
   );
-}
-
-export type ActiveKiloPassByFingerprint = {
-  kiloUserId: string;
-  subscriptionId: string;
-  stripeSubscriptionId: string | null;
-};
-
-export async function findActiveKiloPassByCardFingerprint(
-  fingerprint: string | null | undefined,
-  excludingUserId: string
-): Promise<ActiveKiloPassByFingerprint | null> {
-  if (!fingerprint) return null;
-
-  const otherPaymentMethods = await db
-    .select({
-      userId: payment_methods.user_id,
-    })
-    .from(payment_methods)
-    .where(
-      and(
-        eq(payment_methods.stripe_fingerprint, fingerprint),
-        ne(payment_methods.user_id, excludingUserId)
-      )
-    )
-    .limit(10);
-
-  const otherUserIds = [...new Set(otherPaymentMethods.map(pm => pm.userId))];
-  if (otherUserIds.length === 0) return null;
-
-  const activeSub = await db
-    .select({
-      kiloUserId: kilo_pass_subscriptions.kilo_user_id,
-      id: kilo_pass_subscriptions.id,
-      stripeSubscriptionId: kilo_pass_subscriptions.stripe_subscription_id,
-    })
-    .from(kilo_pass_subscriptions)
-    .where(
-      and(
-        inArray(kilo_pass_subscriptions.kilo_user_id, otherUserIds),
-        isNull(kilo_pass_subscriptions.ended_at),
-        sql`${kilo_pass_subscriptions.status} NOT IN ('canceled', 'unpaid', 'incomplete_expired')`
-      )
-    )
-    .limit(1);
-
-  if (activeSub.length === 0) return null;
-
-  return {
-    kiloUserId: activeSub[0].kiloUserId,
-    subscriptionId: activeSub[0].id,
-    stripeSubscriptionId: activeSub[0].stripeSubscriptionId,
-  };
 }
 
 export type StripeTopupMetadata = {
